Chez Scheme: reduce memory use of pbchunk conversion
The C code generated by the pbchunk conversion is especially sparse,
with very large macro names for simple operations and lots of comments
to reflect source instructions. Generation of that code was formerly
rendering to text in memory, and then splitting into the output files.
Converting to text at the last minute can have a much lower peak
memory use.