raco demod: support keeping syntax objects, macros, and provides
Improve the demodularizer so that it can be used to flatten a library
for use by other modules, and not just flatten a complete program.
Specifically, the improvements enable preservation of syntax objects,
provides, and compile-time code (that implements macros).

The main remaining limitation is submodules: any submodule of a
demodularized module is currently just discarded. Adding submodule
support will will be a next step.

Meanwhile, using `raco demod` is clunky (the output is a ".zo" file
alongside the source) and it does not play nicely with `raco make` or
`raco setup`. Another next step is adding `#lang compiler/demod` to
define a module whose implementation is the flattened form of another
module.