add `compiler/demod`, support submodules via `raco demod`
The `compiler/demod` module language performs the same kind of
demodularization as `raco demod`, but in a way that works with `raco
make`, `raco setup`, so it's much easier to use for demodularizing a
library (as opposed to a program for end users).

The main use of `#lang compiler/demod` is to reduce the startup time
of a library that is organized into many modules, where "many" means
"hundreds". For example, the core of the current Rhombus
implementation is 300+ modules; demodularizing reduces startup time
for a `#lang rhombus "Hello world"` by about 25%. That's not a
dramatic effect, but it removes a penalty for organizing code nicely.