repair `(thread #:keep 'results ...)` for immediately broken thread
The repair simplifies engine startup in a way that seems clearly
simpler and better, and I'm not sure why it was the old way. The old
pattern originated with dc047d89226, which doesn't seem related
otherwise. I'm chalking it up to confusion in the early days of Racket
CS implementation.