FFI: retain converted arguments to a foreign call
When arguments are supplied to a foreign call, the original arguments
have always been retained, which is helpful if the call leads to a
callback that can trigger a GC. If an argument is converted to a
primitive repersentation via a derived C type, however, the converted
form was not reatained. For example, a `_string/immobile` type might
convert a string to an immobile copy with the idea that it will not
only be retained, but it will also stay put; unfortunately, the
immobile copy will not be retained.

The change here is to retain converted arguments as well as the
original. This change can increase register pressure and spilling
slightly for foreign calls in CS, and it also impose a little more
overhead and potentially more allocation (for an unusually large
number of arguments) for BC, but it seems worthwhile to make things
behave more as expected in a layer that is already tricky.